Added my custom wheel centering code
the reason for this is for you to check it out.
It work better for my wheel than the other 2 options but this could be my wheel only

soft spring forever rotates backwards and forwards unless i turn power well down. then it dosnt have the power to actual center correctly

the second optoins just gains speed not reduce, and again if i lower power it dosnt not center properly